backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
firebase-admin
plaid-python
transformers
torch
google-generativeai


